# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Main Bot API",
    "version": "1.0.0",
    "description": "The main bot api allows you to interact with the BuildTheEarth Discord in a secure and limited fashion.\n\nNote that this is a very restricted API and you need your IP to be whitelisted in order for your requests to work.\n\n### Check out its awesome features:\n\n* Fetch roles\n* Update builder roles\n* Send templated messages to users\n* Create a report against an user\n",
    "contact": {}
  },
  "servers": [
    {
      "url": "https://{ip}:{port}",
      "variables": {
        "ip": {
          "default": "demo",
          "description": "Get the IP"
        },
        "port": {
          "default": "8080"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/role": {
      "get": {
        "tags": [
          "website",
          "btbot"
        ],
        "summary": "Gives the Discord roles of an user",
        "description": "Returns the Discord roles of an user",
        "operationId": "roleStatus",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The Discord Snowflake User ID of target",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string",
              "example": "635411595253776385"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roles found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "User ID",
                      "example": "635411595253776385"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "description": "Role name",
                            "example": "Admin"
                          },
                          "id": {
                            "type": "string",
                            "description": "Role ID",
                            "example": "692504956036710440"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Input Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/paramError"
                }
              }
            }
          },
          "404": {
            "description": "User not foud",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFoundError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "website"
        ],
        "summary": "Add or remove roles from a user",
        "description": "Adds or removes roles from a Discord user in both main and staff servers. Uses role keys that map to role IDs in the whitelist.",
        "operationId": "roleModify",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The Discord Snowflake User ID of target",
            "required": true,
            "schema": {
              "type": "string",
              "example": "635411595253776385"
            }
          }
        ],
        "requestBody": {
          "description": "Request Body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["add", "roles"],
                "properties": {
                  "add": {
                    "type": "boolean",
                    "description": "True to add roles, false to remove roles",
                    "example": true
                  },
                  "roles": {
                    "type": "array",
                    "description": "Array of role keys to add or remove (keys must be in whitelist and map to role IDs for main/staff servers)",
                    "items": {
                      "type": "string",
                      "example": "buildteam_owner"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Roles modified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "description": "User ID",
                      "example": "635411595253776385"
                    },
                    "operation": {
                      "type": "string",
                      "description": "Operation performed",
                      "example": "add"
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "main": {
                          "type": "object",
                          "description": "Results for main Discord server",
                          "properties": {
                            "success": {
                              "type": "array",
                              "description": "Role keys successfully modified",
                              "items": {
                                "type": "string",
                                "example": "buildteam_owner"
                              }
                            },
                            "failure": {
                              "type": "array",
                              "description": "Roles that failed to modify",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "roleKey": {
                                    "type": "string"
                                  },
                                  "error": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        },
                        "staff": {
                          "type": "object",
                          "description": "Results for staff Discord server",
                          "properties": {
                            "success": {
                              "type": "array",
                              "description": "Role keys successfully modified",
                              "items": {
                                "type": "string",
                                "example": "buildteam_owner"
                              }
                            },
                            "failure": {
                              "type": "array",
                              "description": "Roles that failed to modify",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "roleKey": {
                                    "type": "string"
                                  },
                                  "error": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Input Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/paramError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Role key not in whitelist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "FORBIDDEN"
                    },
                    "message": {
                      "type": "string",
                      "example": "One or more role keys are not allowed to be modified"
                    },
                    "invalidRoles": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "unknown_role_key"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found in any guild",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/notFoundError"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - Whitelist not configured",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "SERVICE_UNAVAILABLE"
                    },
                    "message": {
                      "type": "string",
                      "example": "Role whitelist is not configured or empty"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "paramError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "MISSING_PARAMETER",
            "description": "Usually missing or invalid param, perhaps WRONG_DATATYPE"
          },
          "message": {
            "type": "string",
            "example": "Missing parameter: id"
          }
        }
      },
      "notFoundError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Not found: user"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": []
}